Bug fix: declare a variable of type uint32_t instead of size_t.
When unpacking an XrlList argument inside XrlAtom::unpack_list(),
the local variable named "nelem" that contains the number of elements
inside an XRL must be defined as uint32_t instead of size_t,
because later the code uses sizeof(nelem) to calculate the number
of consumed bytes. On 64-bit machine, size_t is different from uint32_t,
hence the number of consumed bytes is incorrect.

Bug found on OpenBSD + amd64.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@5536 26435e26-4b25-483a-942d-477236992e84

+++ trunk/xorp/libxipc/xrl_atom.cc
@@ -12,7 +12,7 @@
 // notice is a summary of the XORP LICENSE file; the license in that file is
 // legally binding.
 
-#ident "$XORP: xorp/libxipc/xrl_atom.cc,v 1.16 2005/01/21 01:59:49 atanu Exp $"
+#ident "$XORP: xorp/libxipc/xrl_atom.cc,v 1.17 2005/03/03 07:36:12 pavlin Exp $"
 
 #include <sys/types.h>
 #include <sys/socket.h>
@@ -906,7 +906,7 @@ size_t
 XrlAtom::unpack_list(const uint8_t* buffer, size_t buffer_bytes)
 {
     size_t used = 0;
-    size_t nelem;
+    uint32_t nelem;
 
     if (buffer_bytes < sizeof(nelem)) {
 	return 0;
